Fix PhantomJS issues
